\name{spdyn-package}
\alias{spdyn-package}
\alias{spdyn}
\docType{package}
\title{
Exploratory space-time data analysis
}
\description{
This package implements methdos and measures to analyze longitudinal spatial data 
}
\details{
\tabular{ll}{
Package: \tab spdyn\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-28\cr
License: \tab GPL-3\cr
}
An overview of how to use the package, including the most important functions
}
\author{
Osmar Leandro Loaiza Quintero

Maintainer: Osmar L. Loaiza Q. <olloaizaq at unal.edu.cp>

}
\references{
Restrepo, Franco and Muñoz (2010) Algebra Lineal con Aplicaciones. Universidad Nacional de Colombia.

Rey, Sergio (2001) 'Spatial Empirics for Economic Growth and Convergence', Geographical Analysis, Vol.33, No.3.

}
 
\keyword{ spatial data, transition markov matrices }
\seealso{
Optional links to other man pages, e.g.
\code{\link[<pkg>:<spdep>-package]{<pkg>}} ~~
}
\examples{
#simple examples of the most important functions
}
