\name{spdyn}
\alias{spdyn}
\alias{spdyn}
\docType{package}
\title{
Exploratory space-time data analysis
}
\description{
This package implements methods and measures to analyze longitudinal spatial data. 
}
\details{
\tabular{ll}{
Package: \tab spdyn\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2014-05-28\cr
License: \tab GPL-3\cr
}

}
\author{
Osmar Leandro Loaiza Quintero

\emph{Maintainer}: Osmar L. Loaiza Q. <olloaizaq at unal.edu.co>

}
\references{
Restrepo, Patricia; Franco, Rosa and Muñoz, Luz (2010). \emph{Álgebra Lineal con Aplicaciones}, Universidad Nacional de Colombia, Medellín.

Rey, Sergio (2001). 'Spatial Empirics for Economic Growth and Convergence', \emph{Geographical Analysis}, Vol.33, No.3.
}
 
\keyword{ spatial data, transition markov matrices }

